requirements.txt
pygame-ce==2.5.3
pygbag==0.9.2
